change sp uint16 to uint8

+++ includes/cpu.h
@@ -6,7 +6,7 @@
 /*   By: gabch <gabch@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/22 17:07:32 by gabch             #+#    #+#             */
-/*   Updated: 2026/05/23 20:00:55 by gabch            ###   ########.fr       */
+/*   Updated: 2026/05/26 20:37:22 by gabch            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ typedef struct s_cpu
 	uint8_t		x;
 	uint8_t		y;
 	uint8_t		sr;
-	uint16_t	sp;
+	uint8_t		sp;
 	uint16_t	pc;
 }	t_cpu;
